requirements.txt
Django==3.2.11
pillow==9.0.1
djangorestframework==3.13.1
